Fix and improve makecontext
The original design of makecontext requires each sub fiber to manually call swapcontext() to switch back to the main thread;
This is tedious, and user unfriendly. This commit improves makecontext and fix bugs, so that simply pointing uc_link to main
thread will do the trick.

get_sp() and get_pc() are added to debug.s for debugging