Change proc struct structure information to using i_flags
Previously struct proc uses field IN_USE to check if the slot is in use. However, this is unnecessary as only 1 bit is used in the 32bit
field. Rather i_flags(information flags) is created instead, which can be used using AND operationgs.

Shell now prints helper message for list of available commands