Add times(2) syscall
times(2) is now supported in winix, though cutime and cstime is not
returned. Ipc message now supports unions of different struct.
Billable flags is added to struct proc field i_flags. When this flag
is set, it means the process to which bill_ptr is pointing to can be
billed for each interrupt the current process consumes.