Add errno string list in user lib; Add system debugger tools
System debugger tools, saved in tools/kdbg.py, is an extern script
that is used to debug kernel, or any of the user process.

Given the PC address (usually the address at the the process crashed)
, the script analyse the corresponding binary file, and return the
actual file line number and the file where the crash happened.

e.g. if the kernel crashed at PC = 0x2764
	py3 tools/kdbg.py 0x2764
     if shell crashed at virtual PC = 0x372
	py3 tools/kdbg.py 0x372 sh
     if init crashed at virtual PC = 0x27
	py3 tools/kdbg.py 0x27 init