BUG FIX: search process by pid from user space to kernel space
Process number a.k.a proc_nr is used to uniquely identify each
process in the kernel space, whereas pid is used to identify process
in the user space. When pid is passed from user space to kernel
space, get_proc_by_pid() is used to search process with the
corresponding pid.

In addition, pid is returned to the user space on return of a
system call