Implement sigaction syscall
In particular, instead of building sigreturn assembly call on the stack,
sigreturn is implemented in the user library, and the virtual
address is passed to the kernel when making sigaction syscall.
The kernel will set ra to __sigreturn before invoking the sig
handler.

Fix is_bit_on() bug where the opposite of the result is returned.