Renmae sig_proc() to cause_sig(); rewrite exit_proc()
exit_proc() takes three parameters: process to be exited, exit
status, and signal number which killed the process. This allows
exit_proc() to be a universal function for both exit() and stopped
processes.

kdbg.py is rewritten to display the .data and .bss segment
information given the address provided in the command line argument.